refactor: improve project-info page organization
- Created shared types for project info
- Extracted state management into useProjectInfo hook
- Created ProjectInfoLayout component for UI structure
- Simplified main page component
- Improved type safety with better type definitions
